% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam2disk.R
\name{cam2disk}
\alias{cam2disk}
\title{Copy and Rename with folder structure}
\usage{
cam2disk(
  source = NULL,
  dest = NULL,
  img_format = c("png", "jpg", "gpr"),
  vid_format = c("mp4"),
  names_levs = 3,
  rename = TRUE,
  sep = "-",
  prefix = NULL,
  add_seq = TRUE,
  from_file = NULL,
  to_file = NULL,
  from_daytime = NULL,
  to_daytime = NULL,
  time_format = "\%Y:\%m:\%d \%H:\%M:\%S"
)
}
\arguments{
\item{source}{absolute or relative path to the folder/drive where the files are stored (e.g., an SD )}

\item{dest}{absolute path to where the files need to be copied}

\item{img_format}{character vector (case insensitive) specifying in which formats the image files could be provided. List all formats that are being used in the project. Any format not specified will not be considered. Formats are case insensitive: formats differing only in case will be treated as the same format (i.e., "jpg" == "JPG")}

\item{vid_format}{character vector (case insensitive) specifying in which formats the video files could be provided. List all formats that are being used in the project. Any format not specified will not be considered. Formats are case insensitive: formats differing only in case will be treated as the same format (i.e., "mp4" == "MP4")}

\item{rename}{logical. Default to TRUE means that the copied files will also be renamed.}

\item{sep}{character use to separate the different components of the file name}

\item{prefix}{(optional) a string to be attaced to the beginning of the file name}

\item{add_seq}{logical. If TRUE (default) a sequential number is added at the end of the name, before the format extension, for each format type.}

\item{from_file}{name of the first file of those in the interval to be copied (and renamed). See \link{details}}

\item{to_file}{name of the last file file of those in the interval to be copied (and renamed). See \link{details}}

\item{from_daytime}{timestamp marking the beginning of the time interval that includes the files to be copied on the base of when they were created originally (i.e., when the were created by the device that generated them).}

\item{to_daytime}{timestamp marking the end of the time interval that includes the files to be copied on the base of when they were created originally (i.e., when the were created by the device that generated them).}

\item{time_format}{format in which \code{from_daytime} and \code{to_daytime} are expressed as in \link{strptime}.}

\item{name_levs}{numeric, given the path of \code{dest}, it specify the last N levels of the path to be used in renaming a file (e.g., with dest= c:/lev1/lev2/lev3/lev4 and names_levs=2 then lev3 and lev4 will be used in the name)}
}
\description{
Copy files from a camera and copy them onto another drive while organizing them into a folder hierarchy.
}
\details{
The time selection is based on the variable \code{ModifyDate} from the dataframe obtained using \link[exiftoolr:exif_read]{exiftoolr::exif_read}. This because on development it seemed more reliable than CreatDate (which had unreal values for GRP files).

From_file and to_file could be specified with or without extension. In case the extension is not provided, all files matching the name will be copied.
Selection of file by name has priority over time selection.

The function rely on Exif data from the original file to get the file metadata to work with. In particular, Exif is used to get the time stamp of when the file was first created (i.e., by a device such as camera, or phone or gps)
For files other than those created by divices that provide Exif metadata, the function should still work but only wit name-based selection

#Warning
The function will not overwrite existing file, however no warning that files already exist is give. To be on the safe side erase files in the destination folder before copying.
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
