% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sync-0.2.R
\name{estimate_sync}
\alias{estimate_sync}
\title{Estimate reference points in images sequences}
\usage{
estimate_sync(
  source = NULL,
  ref_data = NULL,
  output_path = NULL,
  shooting_group = "shooting_group",
  cameraID = "cameraID",
  camera_role = "camera_role",
  sequence = "sequence",
  ref_point = "ref_point",
  file_name = "file_name"
)
}
\arguments{
\item{source}{absolute or relative path to the folder/drive where the subfolders with the image files of the individual cameras are stored.}

\item{ref_data}{dataframe with information about reference points in the sequence for Master and initial ref points for slaves.}

\item{output_path}{absolute or relative path to the output file (.csv). It must end with the file name}

\item{shooting_group}{name of the column in the dataset specifying the group of picture taken in the same shooting session}

\item{cameraID}{name of the column in the dataset grouping the picture by camera}

\item{camera_role}{name of the column in the dataset specifying the role of the camera. That is 'master' for the camera used to retrive reference points, 'slave' for the other cameras that need to be synchronized with the 'master'}

\item{sequence}{name of the column in the dataset grouping the images into a continuous sequence within the shooting session (e.g., a transect from start to end)}

\item{ref_point}{name of the column in the dataset specifying if the reference point is the 'start' or the 'end' of the sequence}

\item{file_name}{name of the column in the dataset specifying the file name of the image taken as reference point}
}
\description{
Given a set of reference points in a sequence of images taken at regular intervals (i.e., the file name of an image corresponding to a known scene, such as the beginning or the end of a chunk of images of interest),
the function estimates the file name of corresponding images in a different image sequence taken at the same regular interval.
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
