% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_max_spacing.R
\name{find_max_spacing}
\alias{find_max_spacing}
\title{Find maximum allowed spacing}
\usage{
find_max_spacing(fov_L, fov_R = fov_L, min_camera_overlap, dist)
}
\arguments{
\item{fov_L}{a numeric value/vector expressing FOV in RADIANTS. Either the FOV of all the cameras, or the left camera in a pair if the right one has a different FOV. In this case specify FOV_R.}

\item{fov_R}{a numeric value/vector expressing FOV in RADIANTS. Optional, if not specified it is assument to be equal to fov_L.}

\item{min_camera_overlap}{desired miminum overalp between two adjacent cameras on the rig}

\item{dist}{a numeric value/vector expressing the distance between the camera and the plane of interest (e.g., the bottom, the ground)}
}
\value{
A matrix with named dimensions: columns are minimum overlap, rows are distances from the plane of interest. Attention: remembere that dimnames are characters, if used to subset.
}
\description{
Find maximum allowed spacing between cameras given distance to the plane of interest and the desired minimum percentage of overalp in swath between two adjacent cameras.
}
\examples{

find_max_spacing(fov_L=1.42 , min_camera_overlap= seq(40, 60, 10),dist=seq(70, 100, 10))


}
\seealso{
\link{imaging_strategy}, \link{swath_rig}, \link{FOV}, \link{WID}, \link{overlap_pc_cam}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
