% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fov_functions.R
\name{FOV}
\alias{FOV}
\alias{WID}
\alias{rad2deg}
\title{Camera properties}
\usage{
FOV(C, d)

WID(fov, d)

rad2deg(x)
}
\arguments{
\item{C}{Horizontal distance of the observable extent (either estimated or measured) in the scene. Numeric}

\item{d}{Distance of the camera from the bottom (or plane of interest)}

\item{fov}{Field of view in radiants, estimated.}

\item{x}{an angle expressed in radiants.}
}
\description{
Functions to calculate a camera field of view and footprint on the bottom of an image given a distance to it.
}
\references{
Rende SF, Irving AD, Bacci T, Parlagreco L, Bruno F, De Filippo F, Montefalcone M, Penna M, Trabucco B, Di Mento R, et al. 2015. Advances in micro-cartography: a two-dimensional photo mosaicing technique for seagrass monitoring. Estuarine, Coastal and Shelf Science 167(November 2015): 475–486. Elsevier Ltd. doi: 10.1016/j.ecss.2015.10.029.
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
