% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{imaging_strategy}
\alias{imaging_strategy}
\title{Define the imaging strategy}
\usage{
imaging_strategy(
  fov_L,
  dist,
  cam_spacing,
  min_camera_overlap = NULL,
  n_cams,
  transect_pos = c("side"),
  corridor_width = NULL
)
}
\arguments{
\item{fov_L}{a single numeric value expressing FOV in RADIANTS. Either the FOV of all the cameras, or the left camera in a pair if the right one has a different FOV. In this case specify FOV_R.}

\item{dist}{a single numeric value expressing the distance between the camera and the plane of interest (e.g., the bottom, the ground)}

\item{cam_spacing}{a single numeric value expressing the spacing between two cameras}

\item{min_camera_overlap}{Minimum desired overalp, expressed as a percentage, between frames of the (distal) camera during opposed filming passages. It specifically refers to the overlap in swath of the one single camera that is at the distal end of a rig.}

\item{n_cams}{number of cameras in the rig (all considred having the same properties and being alligned)}

\item{transect_pos}{the posistion on the bottom of the a reference transect (if used) relative to the center of the movement of the swath of the rig. Options are "side" or "center".}

\item{corridor_width}{Distance between two consecutive reference transect lines. Defalult to NULL.}
}
\description{
Help define the imaging strategy depending on the camera and rig properties, distance from the bottom and position of reference transects.
}
\details{
A filming passage consists in one leg of the lawn mower path, so that two adjacent passages are usually filmed moving in opposit directions and are meant to have an overlap when they cross.
When \code{transect_pos} is 'side', the imaging strategy consists in running one passage on one side of a reference transect, and one opposite passage on the other side of the transect taking care of having a deired degree of verlap.
When \code{transect_pos} is 'center', the imaging strategy consists in running one passage on the transect. This strategy is most meaningful if used with corridors.

\code{min_camera_overlap} is the \emph{minimum} desired overlap between opposites passages. It defines the \emph{maximum} distance from the transect when \code{transect_pos} is 'side' to keep to avoid obtaining an overlap \emph{less than} the \code{min_camera_overlap}.
Also \code{min_camera_overlap} is considered the minimum desired overlap between opposites extra passages.
}
\examples{

imaging_strategy(fov_L=1.42, dist=100, cam_spacing=41, transect_pos='side',min_camera_overlap=50, n_cams=3, corridor_width=NULL)
imaging_strategy(fov_L=1.42, dist=100, cam_spacing=41, transect_pos='side',min_camera_overlap=50, n_cams=3, corridor_width=600)
	
imaging_strategy(fov_L=1.42, dist=100, cam_spacing=41, transect_pos='center',min_camera_overlap=50, n_cams=3, corridor_width=NULL)
imaging_strategy(fov_L=1.42, dist=100, cam_spacing=41, transect_pos='center',min_camera_overlap=50, n_cams=3, corridor_width=600)
	
	

}
\seealso{
\link{swath_rig}, \link{overlap_pc_cam}, \link{FOV}, \link{WID}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
