% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_subsample.R
\name{img_subsample}
\alias{img_subsample}
\title{Subsample image sequence}
\usage{
img_subsample(
  source = NULL,
  paired_synced = NULL,
  dest = NULL,
  interval = 1,
  copy = TRUE,
  shooting_group = "shooting_group",
  cameraID = "cameraID",
  sequence = "sequence",
  ...
)
}
\arguments{
\item{source}{absolute or relative path to the folder/drive where the subfolders with the image files of the individual cameras are stored.}

\item{paired_synced}{csv created by \code{pair_synced} with path of synced files.}

\item{dest}{absolute path to where the files need to be copied}

\item{interval}{interval used for subsampling.}

\item{copy}{Logical. If TRUE (Default) it copies the subsampled files in the \code{dest} folder keeping the folder structure. If FALSE, only a .csv list of the subsampled files is saved in \code{dest}.}

\item{shooting_group}{name of the column in the dataset specifying the group of picture taken in the same shooting session}

\item{cameraID}{name of the column in the dataset grouping the picture by camera}

\item{sequence}{name of the column in the dataset grouping the images into a continuous sequence within the shooting session (e.g., a transect from start to end)}

\item{...}{pass arguments to \link{file.copy}}
}
\description{
Subsample a sequence of images by copying files at a specified interval
}
\details{
The 'interval' at which frequency files are selected: if interval = 3 then the 3rd, 6th, 9th, 12th... are selected, along with the first and last file of the sequence.
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
