% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_sync.R
\name{initiate_sync}
\alias{initiate_sync}
\title{Initiate Synchronization dataframe}
\usage{
initiate_sync(output_path = NULL)
}
\arguments{
\item{output_path}{absolute or relative path to the output file (.csv). It must end with the file name!}
}
\description{
Create a template dataframe to save the information of reference points to be used in the synchronization of the images
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
