% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_percentage.R
\name{overlap_pc}
\alias{overlap_pc}
\title{Overalp estimation Functions}
\usage{
overlap_pc(fov_R, fov_L, dist, cam_spacing)
}
\arguments{
\item{fov_R}{a numeric value/vector expressing FOV in RADIANTS}

\item{fov_L}{a numeric value/vector expressing FOV in RADIANTS}

\item{dist}{a numeric value/vector expressing the distance between the camera and the plane of interest (e.g., the bottom, the ground)}

\item{cam_spacing}{a numeric value/vector expressing the spacing between two cameras}
}
\value{
A matrix with named dimensions: columns are camera spacings, rows are distances from the plane of interest. Attention: remembere that dimnames are characters, if used to subset.
}
\description{
Estimate the percentage of overalp in swathe between two adjacent cameras.
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
