% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_percentage.R
\name{overlap_pc_cam}
\alias{overlap_pc_cam}
\title{Overlap estimation Functions}
\usage{
overlap_pc_cam(fov_L, fov_R = fov_L, dist, cam_spacing)
}
\arguments{
\item{fov_L}{a numeric value/vector expressing FOV in RADIANTS. Either the FOV of all the cameras, or the left camera in a pair if the right one has a different FOV. In this case specify FOV_R.}

\item{fov_R}{a numeric value/vector expressing FOV in RADIANTS. Optional, if not specified it is assument to be equal to fov_L.}

\item{dist}{a numeric value/vector expressing the distance between the camera and the plane of interest (e.g., the bottom, the ground)}

\item{cam_spacing}{a numeric value/vector expressing the spacing between two cameras}
}
\value{
A matrix of estimate the percentage of overalp (\%) with named dimensions: columns are camera spacings, rows are distances from the plane of interest. Attention: remembere that dimnames are characters, if used to subset.
}
\description{
Estimate the percentage of overalp in swath between two adjacent cameras.
}
\seealso{
\link{imaging_strategy}, \link{swath_rig}, \link{FOV}, \link{WID}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
