% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_pair.R
\name{pair_synced}
\alias{pair_synced}
\title{Pair syncronized images}
\usage{
pair_synced(
  source = NULL,
  ref_data = NULL,
  use_validated = F,
  output_path = NULL,
  shooting_group = "shooting_group",
  cameraID = "cameraID",
  camera_role = "camera_role",
  sequence = "sequence",
  ref_point = "ref_point",
  file_name = "file_name"
)
}
\arguments{
\item{source}{absolute or relative path to the folder/drive where the subfolders with the image files of the individual cameras are stored.}

\item{ref_data}{dataset with estimated reference points.}

\item{use_validated}{logical. Wheter using validated reference points or not (the default). Not used yet}

\item{output_path}{absolute or relative path to the output file (.csv). It must end with the file name}

\item{shooting_group}{name of the column in the dataset specifying the group of picture taken in the same shooting session}

\item{cameraID}{name of the column in the dataset grouping the picture by camera}

\item{camera_role}{name of the column in the dataset specifying the role of the camera. That is 'master' for the camera used to retrive reference points, 'slave' for the other cameras that need to be synchronized with the 'master'}

\item{sequence}{name of the column in the dataset grouping the images into a continuous sequence within the shooting session (e.g., a transect from start to end)}

\item{ref_point}{name of the column in the dataset specifying if the reference point is the 'start' or the 'end' of the sequence}

\item{file_name}{name of the column in the dataset specifying the file name of the image taken as reference point}
}
\description{
Pair syncronized images
}
\details{
The reference point could be validated or not.
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
