% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swath.R
\name{swath_rig}
\alias{swath_rig}
\title{Swath of a camera rig}
\usage{
swath_rig(fov_L, fov_R = fov_L, dist, cam_spacing, n_cams = NULL)
}
\arguments{
\item{fov_L}{a numeric value/vector expressing FOV in RADIANTS. Either the FOV of all the cameras, or the left camera in a pair if the right one has a different FOV. In this case specify FOV_R.}

\item{fov_R}{a numeric value/vector expressing FOV in RADIANTS. Optional, if not specified it is assument to be equal to fov_L.}

\item{dist}{a numeric value/vector expressing the distance between the camera and the plane of interest (e.g., the bottom, the ground)}

\item{cam_spacing}{a numeric value/vector expressing the spacing between two cameras}

\item{n_cams}{number of cameras in the rig (all considred having the same properties and being alligned)}
}
\value{
A matrix storing swath values for each compbination of dimnames "dist" and "cam_spacing". The matrix also has between camera overlap stored as an attribute
}
\description{
Estimate the swath that will be obtained using a give number of cameras alligned in a rig.
}
\details{
It is supposed that all the cameras have the same FOV, to be specified as \code{fov_L}.
}
\examples{
swath_rig(fov_L=1.352524, fov_R=1.352524,dist=50, cam_spacing=10, n_cams=3)


sw<-swath_rig( 1.352524, 1.352524,dist=seq(50,100, by=10), cam_spacing=seq(10,100, by=10), n_cams=3)
sw

}
\seealso{
\link{imaging_strategy}, \link{overlap_pc_cam}, \link{FOV}, \link{WID}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
