% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swathe.R
\name{swathe}
\alias{swathe}
\title{Swathe}
\usage{
swathe(fov_L, fov_R, dist, cam_spacing, n_cams = NULL)
}
\arguments{
\item{n_cams}{number of cameras in the setup (all considred having the same properties and being alligned)}
}
\value{
A matrix storing swathe values for each compbination of dimnames "dist" and "cam_spacing". The matrix also has between camera overlap stored as an attribute
}
\description{
Estimate the swathe that will be obtained using a give number of cameras alligned in a rig.
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
