% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_finder.R
\name{sync_finder}
\alias{sync_finder}
\title{Exif-based image pair finder}
\usage{
sync_finder(
  source = NULL,
  main_img = NULL,
  img_format = c("png", "jpg", "gpr"),
  sync_ref = NULL,
  timeLapse = 1,
  cleanTemp = T
)
}
\arguments{
\item{source}{absolute or relative path to the folder/drive where the files are stored (e.g., an SD )}

\item{main_img}{character vector of names of the image files (with file extension) that serves as reference to look for paired images.}

\item{img_format}{character vector (case insensitive) specifying in which formats the image files could be provided. List all formats that are being used in the project. Any format not specified will not be considered. Formats are case insensitive: formats differing only in case will be treated as the same format (i.e., "jpg" == "JPG"). NULL will esclude images.}

\item{sync_ref}{character vector specifying the names of the files which were deemed to be representing the same time point.}

\item{timeLapse}{Time Lapse interval (in seconds).}

\item{cleanTemp}{TRUE to save a copy of the exif dataset in a temp file. FALSE to use an existing copy in the temp file.}
}
\description{
Exif-based image pair finder
}
\details{
Files listed in \code{sync_ref} are ideally images that the operator deemed to be synchronous. Most of the time this are not really shot at the same exact time, but they are manually "aligned" or "synchronized", and so they provide a baseline to determine synchronicity of other images.

Use \code{cleanTemp = TRUE} when running the function in a folder for the first time, or when the files in the folder were modified. Use FALSE otherwise to speed up the runtime.
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
