% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rig_swath_overlap.R
\name{rig_swaths_overlap}
\alias{rig_swaths_overlap}
\title{Calculate the overlap between rig swaths}
\usage{
rig_swaths_overlap(swath_width, dist_centers, external_cam_swath)
}
\arguments{
\item{swath_width}{swath of a rig. E.g., as given by \link{swath_rig}}

\item{dist_centers}{distance between the centers of the two swaths}

\item{external_cam_swath}{swath of the external camera of the rig (i.e., the one that will overlap with the other swath). E.g., as given by \link{WID}}
}
\description{
Calculate the overlap between rig swaths at a given distance between swaths centers.
The overlap is expressed relative to the swath of the external camera on the rig: if >= 1 then the external camera is completly overalpped (not necessairily by only one camera on the opposite rig swath).
}
