% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_filter}
\alias{setup_filter}
\title{Filtering a swath matrix}
\usage{
setup_filter(matrix, cond)
}
\arguments{
\item{matrix}{a matrix produced by \link{swath_rig}.}

\item{cond}{a string specifying the filtering conditions. Important: the string should contain the name of the object passed to \code{matrix}}
}
\value{
A named list containing the filtered matrix (i.e., only showing the result meeting the criteria) and a dataframe of the indices (i.e., distance from the bottom and camera spacing) associated to the selected criteria.
}
\description{
This function help visualising which combination of \emph{distance from the bottom} and \emph{camera spacing} provide a desired range of swaths
}
\examples{

setup_filter(sw, cond='sw>150 & sw<200')



}
\seealso{
\link{swath_rig}, \link{overlap_pc_cam}, \link{imaging_strategy}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
